Conflicts:
	android/host-source/source/project/src/MoaiView.java
	eclipse/android-project/moai-target
	eclipse/libmoai/jni/Android.mk
	eclipse/libmoai/jni/libbox2d.mk
	eclipse/libmoai/jni/libmoaicore.mk
	eclipse/libmoai/jni/libmoaiext-untz.mk
	eclipse/libmoai/jni/libuslscore.mk
	eclipse/libmoai/jni/libuslsext.mk
	eclipse/libmoai/jni/libzipfs.mk
	eclipse/libmoai/jni/moaiJNI.cpp
	eclipse/packager/moai-build.xml
	xcode/libmoai/libmoai.xcodeproj/project.pbxproj

+++ src/aku/AKU-iphone.mm
@@ -115,7 +115,7 @@ void AKUIphoneInit ( UIApplication* application ) {
 	MOAIEnvironment::Get ().SetDevModel ( [[ UIDevice currentDevice ].model UTF8String ] );
 
 	CGRect screenRect = [[UIScreen mainScreen] bounds];
-	MOAIEnvironment::Get ().SetScreenSize( screenRect.size.width, screenRect.size.height );
+	MOAIEnvironment::Get ().SetScreenSize( ( int ) screenRect.size.width, ( int ) screenRect.size.height );
 
 	if ([[ UIScreen mainScreen ] scale ] == 2.0 ) {
 		//this is retina
